+++ src/components/Editor/Project/Project.js
@@ -5,6 +5,7 @@ import "react-tabs/style/react-tabs.css";
 import "react-toastify/dist/ReactToastify.css";
 import { useContainerQuery } from "react-container-query";
 import classnames from "classnames";
+import { useMediaQuery } from "react-responsive";
 
 import "./Project.scss";
 import Output from "../Output/Output";
@@ -29,6 +30,8 @@ import EditorInput from "../EditorInput/EditorInput";
 import NewFileModal from "../../Modals/NewFileModal";
 import ResizableWithHandle from "../../../utils/ResizableWithHandle";
 import { projContainer } from "../../../utils/containerQueries";
+import MobileProject from "../../Mobile/MobileProject/MobileProject";
+import { MOBILE_MEDIA_QUERY } from "../../../utils/mediaQueryBreakpoints";
 
 const Project = (props) => {
   const dispatch = useDispatch();
@@ -121,6 +124,8 @@ const Project = (props) => {
   const [defaultHeight, setDefaultHeight] = useState("auto");
   const [maxWidth, setMaxWidth] = useState("100%");
   const [handleDirection, setHandleDirection] = useState("right");
+  const isMobile = useMediaQuery({ query: MOBILE_MEDIA_QUERY });
+  console.log(MOBILE_MEDIA_QUERY);
 
   useMemo(() => {
     const isDesktop = params["width-larger-than-880"];
@@ -133,31 +138,35 @@ const Project = (props) => {
 
   return (
     <div className="proj">
-      <div
-        ref={containerRef}
-        className={classnames("proj-container", {
-          "proj-container--wc": forWebComponent,
-        })}
-      >
-        {!forWebComponent ? <Sidebar /> : null}
-        <div className="project-wrapper">
-          {!forWebComponent ? <ProjectBar /> : null}
-          <div className="proj-editor-wrapper">
-            <ResizableWithHandle
-              data-testid="proj-editor-container"
-              className="proj-editor-container"
-              defaultWidth={defaultWidth}
-              defaultHeight={defaultHeight}
-              handleDirection={handleDirection}
-              minWidth="25%"
-              maxWidth={maxWidth}
-            >
-              <EditorInput />
-            </ResizableWithHandle>
-            <Output />
+      {isMobile ? (
+        <MobileProject />
+      ) : (
+        <div
+          ref={containerRef}
+          className={classnames("proj-container", {
+            "proj-container--wc": forWebComponent,
+          })}
+        >
+          {!forWebComponent ? <Sidebar /> : null}
+          <div className="project-wrapper">
+            {!forWebComponent ? <ProjectBar /> : null}
+            <div className="proj-editor-wrapper">
+              <ResizableWithHandle
+                data-testid="proj-editor-container"
+                className="proj-editor-container"
+                defaultWidth={defaultWidth}
+                defaultHeight={defaultHeight}
+                handleDirection={handleDirection}
+                minWidth="25%"
+                maxWidth={maxWidth}
+              >
+                <EditorInput />
+              </ResizableWithHandle>
+              <Output />
+            </div>
           </div>
         </div>
-      </div>
+      )}
       {newFileModalShowing ? <NewFileModal /> : null}
       {renameFileModalShowing && modals.renameFile ? <RenameFile /> : null}
       {notFoundModalShowing ? <NotFoundModal /> : null}