+++ CHANGELOG.md
@@ -4,12 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Mobile navigation between code and output (#615)
+
+### Changed
+
+- Position of run button on mobile (#615)
+
 ## [0.18.1] - 2023-08-03
 
 ### Fixed
 
 - Fix list of the projects (#614)
-
 
 ### Changed